refactor: reorganizacion de modulos y refactor a relaciones polimorficas

+++ app/Filament/Resources/Compras/DevolucionCompra/Pages/ViewDevolucionCompra.php
@@ -34,7 +34,7 @@ class ViewDevolucionCompra extends ViewRecord
                 ->modalDescription('Al confirmar, se retirará el stock físico del inventario (registrando movimientos de tipo DEVOLUCION_PROVEEDOR) y se liberará el saldo de la Orden de Compra para futuras recepciones. Esta acción no se puede deshacer.')
                 ->action(function () {
                     try {
-                        app(DevolverMercanciaProveedor::class)->execute($this->record, auth()->id());
+                        app(DevolverMercanciaProveedor::class)->execute($this->record, (int) auth()->id());
 
                         Notification::make()
                             ->title('Devolución Confirmada')
